webrtc: isolate PyNvVideoCodec in a dedicated nvenc module
Split PyNvHardwareEncoder and its ABGR-frame / NAL-scan helpers out of
encoders.py into a sibling nvenc.py that owns the top-level
PyNvVideoCodec import. encoders.py no longer touches PyNvVideoCodec at
module load; select_encoder probes availability via importlib.util.find_spec
(no side effects) and imports nvenc only when a hardware backend is
actually about to be constructed.

The Lingbot WebRTC path resolves to DefaultRTCEncoder in the base
manager and never enters the hardware branch, so its process no longer
loads PyNvVideoCodec at all — sidestepping a silent early-exit observed
during the Lingbot server launch when PyNvVideoCodec was importable.

Tests updated to patch _pynvvideocodec_installed and inject a fake
PyNvVideoCodec into sys.modules so nvenc's top-level import binds to
the mock. The GPU smoke test's imports follow the split.
